Add files via upload
The ChatBot program is a simple conversational AI implemented in Java. It interacts with users, responding to their inputs with predefined greetings and farewells. Users can ask questions or engage in casual conversation with the ChatBot.
Greeting Responses: The ChatBot responds with a random greeting ("Hi!", "Hello!", "Hey there!", "Nice to meet you!") when the user inputs "hello" or "hi".
Farewell Responses: If the user inputs "bye",  with a random farewell message ("Goodbye!", "See you later!", "Take care!", "Goodbye! Have a nice day!").